[Decomp][Spell] Cohesion split: extract target validation to SpellCheckTarget.cpp
