+++ src/game/WorldHandlers/SpellCheckTarget.cpp
@@ -1,0 +1,312 @@
+/**
+ * MaNGOS is a full featured server for World of Warcraft, supporting
+ * the following clients: 1.12.x, 2.4.3, 3.3.5a, 4.3.4a and 5.4.8
+ *
+ * Copyright (C) 2005-2025 MaNGOS <https://www.getmangos.eu>
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ *
+ * World of Warcraft, and all World of Warcraft or Warcraft art, images,
+ * and lore are copyrighted by Blizzard Entertainment, Inc.
+ */
+
+/**
+ * @file Spell.cpp
+ * @brief Spell casting and effect implementation
+ *
+ * This file implements the Spell class which handles spell casting:
+ * - Spell validation and casting requirements
+ * - Spell effect execution (damage, healing, summon, etc.)
+ * - Spell targeting and area effects
+ * - Spell cooldowns and resource costs
+ * - Spell interruption and pushback
+ * - Spell aura application
+ * - Spell hit/miss calculations
+ *
+ * Spells are the primary combat mechanic in WoW, encompassing
+ * abilities, talents, and item effects.
+ *
+ * @see Spell for the spell class
+ * @see SpellAura for spell auras
+ * @see SpellMgr for spell management
+ */
+
+#include "Spell.h"
+#include "Database/DatabaseEnv.h"
+#include "WorldPacket.h"
+#include "WorldSession.h"
+#include "GridNotifiers.h"
+#include "GridNotifiersImpl.h"
+#include "Opcodes.h"
+#include "Log.h"
+#include "UpdateMask.h"
+#include "World.h"
+#include "ObjectMgr.h"
+#include "SpellMgr.h"
+#include "Player.h"
+#include "Pet.h"
+#include "Unit.h"
+#include "DynamicObject.h"
+#include "Group.h"
+#include "UpdateData.h"
+#include "MapManager.h"
+#include "ObjectAccessor.h"
+#include "CellImpl.h"
+#include "Policies/Singleton.h"
+#include "SharedDefines.h"
+#include "LootMgr.h"
+#include "VMapFactory.h"
+#include "BattleGround/BattleGround.h"
+#include "Util.h"
+#include "Chat.h"
+#include "DB2Stores.h"
+#include "SQLStorages.h"
+#include "Vehicle.h"
+#include "TemporarySummon.h"
+#include "SQLStorages.h"
+#include "DisableMgr.h"
+#ifdef ENABLE_ELUNA
+#include "LuaEngine.h"
+#endif /* ENABLE_ELUNA */
+
+/**
+ * @brief Checks whether a target matches the spell's creature type restrictions.
+ *
+ * @param target The target being validated.
+ * @return True if the target type is allowed; otherwise, false.
+ */
+bool Spell::CheckTargetCreatureType(Unit* target) const
+{
+    uint32 spellCreatureTargetMask = m_spellInfo->GetTargetCreatureType();
+
+    // Curse of Doom: not find another way to fix spell target check :/
+    if (m_spellInfo->GetSpellFamilyName() == SPELLFAMILY_WARLOCK && m_spellInfo->GetCategory() == 1179)
+    {
+        // not allow cast at player
+        if (target->GetTypeId() == TYPEID_PLAYER)
+        {
+            return false;
+        }
+
+        spellCreatureTargetMask = 0x7FF;
+    }
+
+    // Dismiss Pet, Taming Lesson and Control Robot skipped
+    if (m_spellInfo->Id == 2641 || m_spellInfo->Id == 23356 || m_spellInfo->Id == 30009)
+    {
+        spellCreatureTargetMask =  0;
+    }
+
+    if (spellCreatureTargetMask)
+    {
+        uint32 TargetCreatureType = target->GetCreatureTypeMask();
+
+        return !TargetCreatureType || (spellCreatureTargetMask & TargetCreatureType);
+    }
+    return true;
+}
+
+/**
+ * @brief Gets the current spell container slot used by this spell.
+ *
+ * @return The current spell container type.
+ */
+CurrentSpellTypes Spell::GetCurrentContainer()
+{
+    if (IsNextMeleeSwingSpell())
+    {
+        return (CURRENT_MELEE_SPELL);
+    }
+    else if (IsAutoRepeat())
+    {
+        return (CURRENT_AUTOREPEAT_SPELL);
+    }
+    else if (IsChanneledSpell(m_spellInfo))
+    {
+        return (CURRENT_CHANNELED_SPELL);
+    }
+    else
+    {
+        return (CURRENT_GENERIC_SPELL);
+    }
+}
+
+/**
+ * @brief Validates whether a candidate target is acceptable for a specific effect.
+ *
+ * @param target The target being checked.
+ * @param eff The effect index being validated.
+ * @return True if the target is valid for the effect; otherwise, false.
+ */
+bool Spell::CheckTarget(Unit* target, SpellEffectIndex eff)
+{
+    SpellEffectEntry const* spellEffect = m_spellInfo->GetSpellEffect(eff);
+    if (!spellEffect)
+    {
+        return false;
+    }
+
+    // Check targets for creature type mask and remove not appropriate (skip explicit self target case, maybe need other explicit targets)
+    if (spellEffect->EffectImplicitTargetA != TARGET_SELF )
+    {
+        if (!CheckTargetCreatureType(target))
+        {
+            return false;
+        }
+    }
+
+    // Check Aura spell req (need for AoE spells)
+    SpellAuraRestrictionsEntry const* auraRestrictions = m_spellInfo->GetSpellAuraRestrictions();
+    if (auraRestrictions)
+    {
+        if (auraRestrictions->targetAuraSpell && !target->HasAura(auraRestrictions->targetAuraSpell))
+        {
+            return false;
+        }
+        if (auraRestrictions->excludeTargetAuraSpell && target->HasAura(auraRestrictions->excludeTargetAuraSpell))
+        {
+            return false;
+        }
+    }
+
+    // Check targets for not_selectable unit flag and remove
+    // A player can cast spells on his pet (or other controlled unit) though in any state
+    if (target != m_caster && target->GetCharmerOrOwnerGuid() != m_caster->GetObjectGuid())
+    {
+        // any unattackable target skipped
+        if (target->HasFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_NON_ATTACKABLE))
+        {
+            return false;
+        }
+
+        // unselectable targets skipped in all cases except TARGET_SCRIPT targeting or vehicle passengers
+        // in case TARGET_SCRIPT target selected by server always and can't be cheated
+        if ((!m_IsTriggeredSpell || target != m_targets.getUnitTarget()) &&
+            target->HasFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_NOT_SELECTABLE) &&
+            (!target->GetTransportInfo() || (target->GetTransportInfo() &&
+            !((Unit*)target->GetTransportInfo()->GetTransport())->IsVehicle())) &&
+            spellEffect->EffectImplicitTargetA != TARGET_SCRIPT &&
+            spellEffect->EffectImplicitTargetB != TARGET_SCRIPT &&
+            spellEffect->EffectImplicitTargetA != TARGET_AREAEFFECT_INSTANT &&
+            spellEffect->EffectImplicitTargetB != TARGET_AREAEFFECT_INSTANT &&
+            spellEffect->EffectImplicitTargetA != TARGET_AREAEFFECT_CUSTOM &&
+            spellEffect->EffectImplicitTargetB != TARGET_AREAEFFECT_CUSTOM &&
+            spellEffect->EffectImplicitTargetA != TARGET_NARROW_FRONTAL_CONE &&
+            spellEffect->EffectImplicitTargetB != TARGET_NARROW_FRONTAL_CONE &&
+            spellEffect->EffectImplicitTargetA != TARGET_NARROW_FRONTAL_CONE_2 &&
+            spellEffect->EffectImplicitTargetB != TARGET_NARROW_FRONTAL_CONE_2)
+            {
+                return false;
+            }
+    }
+
+    // Check player targets and remove if in GM mode or GM invisibility (for not self casting case)
+    if (target != m_caster && target->GetTypeId() == TYPEID_PLAYER)
+    {
+        if (((Player*)target)->GetVisibility() == VISIBILITY_OFF)
+        {
+            return false;
+        }
+
+        if (((Player*)target)->isGameMaster() && !IsPositiveSpell(m_spellInfo->Id))
+        {
+            return false;
+        }
+    }
+
+    // Check targets for LOS visibility (except spells without range limitations )
+    if (!DisableMgr::IsDisabledFor(DISABLE_TYPE_SPELL, m_spellInfo->Id, NULL, SPELL_ATTR_EX2_IGNORE_LOS))
+    {
+        switch(spellEffect->Effect)
+        {
+        case SPELL_EFFECT_SUMMON_PLAYER:                    // from anywhere
+            break;
+        case SPELL_EFFECT_DUMMY:
+            if (m_spellInfo->Id != 20577)                   // Cannibalize
+            {
+                break;
+            }
+            // fall through
+        case SPELL_EFFECT_RESURRECT_NEW:
+            // player far away, maybe his corpse near?
+            if (target != m_caster && !m_spellInfo->HasAttribute(SPELL_ATTR_EX2_IGNORE_LOS) && !target->IsWithinLOSInMap(m_caster))
+            {
+                if (!m_targets.getCorpseTargetGuid())
+                {
+                    return false;
+                }
+
+                Corpse* corpse = m_caster->GetMap()->GetCorpse(m_targets.getCorpseTargetGuid());
+                if (!corpse)
+                {
+                    return false;
+                }
+
+                if (target->GetObjectGuid() != corpse->GetOwnerGuid())
+                {
+                    return false;
+                }
+
+                if (!m_spellInfo->HasAttribute(SPELL_ATTR_EX2_IGNORE_LOS) && !corpse->IsWithinLOSInMap(m_caster))
+                {
+                    return false;
+                }
+            }
+
+            // all ok by some way or another, skip normal check
+            break;
+        default:                                            // normal case
+            // Get GO cast coordinates if original caster -> GO
+            if (target != m_caster)
+            {
+                if (WorldObject* caster = GetCastingObject())
+                {
+                    if (!m_spellInfo->HasAttribute(SPELL_ATTR_EX2_IGNORE_LOS) && !target->IsWithinLOSInMap(caster))
+                    {
+                        return false;
+                    }
+                }
+            }
+            break;
+        }
+    }
+
+    if (target->GetTypeId() != TYPEID_PLAYER && m_spellInfo->HasAttribute(SPELL_ATTR_EX3_TARGET_ONLY_PLAYER)
+        && spellEffect->EffectImplicitTargetA != TARGET_SCRIPT && spellEffect->EffectImplicitTargetA != TARGET_SELF)
+    {
+        return false;
+    }
+
+    switch (m_spellInfo->Id)
+    {
+        case 37433:                                         // Spout (The Lurker Below), only players affected if its not in water
+            if (target->GetTypeId() != TYPEID_PLAYER || target->IsInWater())
+            {
+                return false;
+            }
+            break;
+        case 68921:                                         // Soulstorm (FoS), only targets farer than 10 away
+        case 69049:                                         // Soulstorm            - = -
+            if (m_caster->IsWithinDist(target, 10.0f, false))
+            {
+                return false;
+            }
+            break;
+        default:
+            break;
+    }
+
+    return true;
+}
